requirements.txt
autopep8==1.6.0
cachelib==0.4.1
click==8.0.3
colorama==0.4.4
dlib==19.22.1
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Session==0.4.0
Flask-WTF==1.0.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
opencv-contrib-python==4.5.4.60
opencv-python==3.4.16.59
Pillow==9.0.0
pycodestyle==2.8.0
scipy==1.7.3
toml==0.10.2
visitor==0.1.3
waitress==2.0.0
Werkzeug==2.0.2
WTForms==3.0.1